Renaming SwiftStrings
Renaming SwiftStrings
