+++ Sources/KSCrash/swift/Raygun_SwiftStrings.h
